+++ src/WebJobs.Extensions.DurableTask/ContextInterfaces/IDurableOrchestrationContext.cs
@@ -114,8 +114,9 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
         /// <param name="method">HttpMethod used for api call.</param>
         /// <param name="uri">uri used to make the HTTP call.</param>
         /// <param name="content">Content passed in the HTTP request.</param>
+        /// <param name="retryOptions">The retry option for the HTTP task.</param>
         /// <returns>A <see cref="Task{DurableHttpResponse}"/>Result of the HTTP call.</returns>
-        Task<DurableHttpResponse> CallHttpAsync(HttpMethod method, Uri uri, string content = null);
+        Task<DurableHttpResponse> CallHttpAsync(HttpMethod method, Uri uri, string content = null, HttpRetryOptions retryOptions = null);
 
         /// <summary>
         /// Makes an HTTP call using the information in the DurableHttpRequest.