+++ test/Common/DurableTaskEndToEndTests.cs
@@ -209,6 +209,55 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
 
 #if !FUNCTIONS_V1
         /// <summary>
+        /// By simulating the appropiate environment variables for Linux Consumption,
+        /// this test checks that we are emitting logs from DurableTask.AzureStorage
+        /// and reading the DurabilityProvider's EventSourceName property correctly.
+        /// </summary>
+        [Fact]
+        [Trait("Category", TestHelpers.DefaultTestCategory)]
+        public async Task AzureStorageEmittingLogsWithEventSourceName()
+        {
+            var prefix = "MS_DURABLE_FUNCTION_EVENTS_LOGS";
+            string orchestratorName = nameof(TestOrchestrations.SayHelloInline);
+
+            // To capture console output in a StringWritter
+            using (StringWriter sw = new StringWriter())
+            {
+                // Set console to write to StringWritter
+                Console.SetOut(sw);
+
+                // Simulate enviroment variables indicating linux consumption
+                var nameResolver = new SimpleNameResolver(new Dictionary<string, string>()
+                {
+                    { "CONTAINER_NAME", "val1" },
+                    { "WEBSITE_STAMP_DEPLOYMENT_ID", "val3" },
+                    { "WEBSITE_HOME_STAMPNAME", "val4" },
+                    { "FUNCTIONS_WORKER_RUNTIME", "python" },
+                });
+
+                // Run trivial orchestrator
+                using (var host = this.testHelper.GetJobHost(
+                    nameResolver: nameResolver,
+                    testName: "FiltersVerboseLogsByDefault",
+                    enableExtendedSessions: false,
+                    storageProviderType: "azure_storage"))
+                {
+                    await host.StartAsync();
+                    var client = await host.StartOrchestratorAsync(orchestratorName, input: "World", this.output);
+                    var status = await client.WaitForCompletionAsync(this.output);
+                    await host.StopAsync();
+                }
+
+                string consoleOutput = sw.ToString();
+
+                // Validate that the JSON has DurableTask-AzureStorage fields
+                string[] lines = consoleOutput.Split('\n');
+                var azureStorageLogLines = lines.Where(l => l.Contains("DurableTask-AzureStorage") && l.StartsWith(prefix));
+                Assert.NotEmpty(azureStorageLogLines);
+            }
+        }
+
+        /// <summary>
         /// By simulating the appropiate enviorment variables for Linux Consumption,
         /// this test checks that we are writing our JSON logs to the console. It does not
         /// verify the contents of the JSON logs themselves (expensive) but instead checks that,
@@ -233,6 +282,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
                     { "CONTAINER_NAME", "val1" },
                     { "WEBSITE_STAMP_DEPLOYMENT_ID", "val3" },
                     { "WEBSITE_HOME_STAMPNAME", "val4" },
+                    { "FUNCTIONS_WORKER_RUNTIME", "powershell" },
                 });
 
                 // Run trivial orchestrator
@@ -289,6 +339,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
             {
                 { "WEBSITE_INSTANCE_ID", "val1" },
                 { "FUNCTIONS_LOGS_MOUNT_PATH", "val2" },
+                { "FUNCTIONS_WORKER_RUNTIME", "python" },
             });
 
             // Run trivial orchestrator
@@ -333,6 +384,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
                     { "CONTAINER_NAME", "val1" },
                     { "WEBSITE_STAMP_DEPLOYMENT_ID", "val3" },
                     { "WEBSITE_HOME_STAMPNAME", "val4" },
+                    { "FUNCTIONS_WORKER_RUNTIME", "python" },
                 });
 
                 // Run trivial orchestrator
@@ -398,6 +450,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
                     { "CONTAINER_NAME", "val1" },
                     { "WEBSITE_STAMP_DEPLOYMENT_ID", "val3" },
                     { "WEBSITE_HOME_STAMPNAME", "val4" },
+                    { "FUNCTIONS_WORKER_RUNTIME", "python" },
                 });
 
                 // Run trivial orchestrator
@@ -464,6 +517,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
             {
                 { "WEBSITE_INSTANCE_ID", "val1" },
                 { "FUNCTIONS_LOGS_MOUNT_PATH", "val2" },
+                { "FUNCTIONS_WORKER_RUNTIME", "python" },
             });
 
             // Run trivial orchestrator
@@ -522,6 +576,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
                 { "FUNCTIONS_LOGS_MOUNT_PATH", "val2" },
                 { "WEBSITE_STAMP_DEPLOYMENT_ID", "val3" },
                 { "WEBSITE_HOME_STAMPNAME", "val4" },
+                { "FUNCTIONS_WORKER_RUNTIME", "python" },
             });
 
             // Run trivial orchestrator
@@ -2853,6 +2908,36 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
 
                 // there should always be some batching going on
                 Assert.True(numBatches < numIterations);
+
+                await host.StopAsync();
+            }
+        }
+
+        /// <summary>
+        /// End-to-end test which validates batching of entity signals.
+        /// </summary>
+        [Theory]
+        [Trait("Category", TestHelpers.DefaultTestCategory)]
+        [InlineData(true)]
+        [InlineData(false)]
+        public async Task DurableEntity_NonexistentEntity(bool extendedSessions)
+        {
+            string[] orchestratorFunctionNames =
+            {
+                nameof(TestOrchestrations.NonexistentEntity),
+            };
+
+            using (var host = this.testHelper.GetJobHost(
+                nameof(this.DurableEntity_NonexistentEntity),
+                extendedSessions))
+            {
+                await host.StartAsync();
+
+                var client = await host.StartOrchestratorAsync(orchestratorFunctionNames[0], null, this.output);
+                var status = await client.WaitForCompletionAsync(this.output);
+
+                Assert.Equal(OrchestrationRuntimeStatus.Completed, status?.RuntimeStatus);
+                Assert.Equal("ok", status?.Output);
 
                 await host.StopAsync();
             }